+++ docs/source/releasenotes.rst
@@ -5,6 +5,8 @@ Upcoming release
 ----------------
 
 - New library: Cloud.Azure
+- Tables: Added keywords ``Get table slice`` and ``Rename table columns``
+- Excel: Files: Fixed various issues with appending data to empty worksheet
 
 1.0.1
 -----